improve indentation of `proc` bodies
A `proc` may not only appear in top-level position but also as local
variable (function literal) and even as parameter within
parentheses. Thus, being within parentheses is not sufficient to
decide the correct indenation. This change use the new function
`nimrod-util-backward-stmt` to identify the defining keyword of a
statement. If the keyword is proc, the indentation is adapted
accordingly.