Use nim-nav-beginning-of-statement function instead
The removed `nim-nav-beginning-of-block` had problem at macro
indentation test.

    # macro
    dumpTree:
      foo:
      x = "bar" # <- failed to indent this line

Also, this `nim-nav-beginning-of-statement` function covers that
general block's logic, so I removed duplicated function.

As additional feature, this function checks string-face or comment-face
and if it is that face, jump over starting point of the string or
comment. (so much stable than I made)